refactor(internal/cli): move LookupCommand
Move LookupCommand from internal/librarian to internal/cli, so that
command execution logic is in one place.

This will also make it easier to refactor the CLI framework in a follow
up PR.
