+++ internal/librarian/librarian.go
@@ -52,7 +52,7 @@ func Run(ctx context.Context, arg ...string) error {
 		CmdLibrarian.Flags.Usage()
 		return nil
 	}
-	cmd, arg, err := lookupCommand(CmdLibrarian, arg)
+	cmd, arg, err := cli.LookupCommand(CmdLibrarian, arg)
 	if err != nil {
 		return err
 	}
@@ -78,30 +78,6 @@ func Run(ctx context.Context, arg ...string) error {
 		return fmt.Errorf("failed to validate config: %s", err)
 	}
 	return cmd.Run(ctx, cmd.Config)
-}
-
-// lookupCommand recursively looks up the command specified by the given arguments.
-// It returns the command, the remaining arguments, and an error if the command
-// is not found.
-func lookupCommand(cmd *cli.Command, args []string) (*cli.Command, []string, error) {
-	if len(args) == 0 {
-		return cmd, nil, nil
-	}
-	subcommand, err := cmd.Lookup(args[0])
-	if err != nil {
-		cmd.Flags.Usage()
-		return nil, nil, err
-	}
-	// If the next argument matches a potential flag (first char is `-`), parse the
-	// remaining arguments as flags. Check if argument is a flag before calling
-	// `lookupCommand` again to avoid flags from being treated as subcommands.
-	if len(args) > 1 && args[1][0] == '-' {
-		return subcommand, args[1:], nil
-	}
-	if len(subcommand.Commands) > 0 {
-		return lookupCommand(subcommand, args[1:])
-	}
-	return subcommand, args[1:], nil
 }
 
 // GitHubClient is an abstraction over the GitHub client.
